+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 # CHANGELOG
 
+1 October 2024
+- Created a Register page
+- Added a link to Register page in Login page
+
+30 September 2024
+- Implemented login mechanism in backend
+- Logging in now redirects to the Dashboard
+
+12 Aug 2024
+- implemented react router
+
 17 July 2024
 - initial commit